Fixes issue 3 (keep-duration > 2 weeks causes feeds to not show new entries).
git-svn-id: https://sparserss.googlecode.com/svn/trunk@20 63f75df1-c233-a9e8-a111-1e0e1c3d5386

+++ src/de/shandschuh/sparserss/handler/RSSHandler.java
@@ -126,7 +126,7 @@ public class RSSHandler extends DefaultHandler {
 	private Date keepDateBorder;
 	
 	public RSSHandler(Context context) {
-		KEEP_TIME = Integer.parseInt(PreferenceManager.getDefaultSharedPreferences(context).getString(Strings.SETTINGS_KEEPTIME, "2"))*86400000;
+		KEEP_TIME = Long.parseLong(PreferenceManager.getDefaultSharedPreferences(context).getString(Strings.SETTINGS_KEEPTIME, "2"))*86400000;
 		keepDateBorder = new Date(System.currentTimeMillis()-KEEP_TIME);
 		this.context = context;
 	}
